rename branch as ga and intergrate cmake
