+++ src/test_ga.cpp
@@ -1,0 +1,5 @@
+#include "genetic_algorithm/base_ga.h"
+
+int main() {
+
+}
